ingest_covid_data/requirements.txt
google-cloud-storage==3.5.0
requests==2.32.5
python-dotenv==1.2.1
